Add latest tag to each FBC image
Adding the `latest` tag to each FBC image makes it easier to fetch the
latest image for each FBC, which we can use to find the correct digest
without having to dig for it manually in Konflux or Quay.

The additional tag is only `latest`, instead of doing something like
`fio-fbc-4-12-latest` since the image repository is already specific to
each OpenShift version and the application, meaning we can keep the
latest tag simple.