Make the update.sh script more idempotent
Previously, running this script twice produced conflicting/duplicate
entries in the catalog.

This commit adds a step to find any existing catalog entries for the one
we're adding and removes them before generating the new catalog, making
it safer to use to update catalogs when we have new bundle images for a
release.

It also relies on a branch tag for the bundle instead of an explicit
digest, making it easier to update catalogs by always fetching the
latest bundle for the branch (and we don't have to update it manually).